+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `isEmail` method
 - Add `fileName` method
 - Add `safeJSONParse` method
+- Add `md5Hash` method
 
 ## 0.0.4 (2024-10-02)